new emulator api
