--- csrc/emulator_api.h
@@ -1,1 +1,0 @@
-../src/main/resources/emulator_api.h
