--- csrc/emulator_mod.h
@@ -1,1 +1,0 @@
-../src/main/resources/emulator_mod.h
